fix for riktig urldirigering til sted
