+++ src/Kart/LeafletTangram/Leaflet.js
@@ -13,6 +13,7 @@ import {
 } from "@material-ui/icons";
 import "style/Kart.scss";
 import updateMarkerPosition from "./LeafletActions/updateMarkerPosition";
+import getLokalitetUrl from "AppSettings/AppFunksjoner/getLokalitetUrl";
 // -- LEAFLET: Fix Leaflet's icon paths for Webpack --
 // See here: https://github.com/PaulLeCam/react-leaflet/issues/255
 // Used in conjunction with url-loader.
@@ -189,18 +190,8 @@ class LeafletTangram extends React.Component {
       if (!data) {
         return null;
       }
-      let url = "";
-      if (data.fylke || data.kommune) {
-        url = "/" + data.kommune.url + "/lokalitet?lng=" + lng + "&lat=" + lat;
-      } else {
-        url = "/Natur_i_Norge/lokalitet?lng=" + lng + "&lat=" + lat;
-      }
-      url = url.replace(/ /g, "_");
-      if (url.substring(0, 2) === "//") {
-        url = url.substring(1);
-      }
+      let url = getLokalitetUrl(lat, lng, data);
       updateMarkerPosition(e, this, header_shift);
-
       this.setState({
         buttonUrl: url,
         data: data,
